Create userDataActions
