+++ package.json
@@ -110,7 +110,7 @@
         "electron-devtools-installer": "2.2.4",
         "electron-is-dev": "1.1.0",
         "mini-css-extract-plugin": "0.8.0",
-        "pc-nrfconnect-shared": "github:NordicSemiconductor/pc-nrfconnect-shared#feature/user-data",
+        "pc-nrfconnect-shared": "github:NordicSemiconductor/pc-nrfconnect-shared#master",
         "spectron": "7.0.0",
         "tar": "4.4.10",
         "util": "0.12.0",
