+++ src/launcher/index.js
@@ -50,7 +50,7 @@ import rootReducer from './reducers';
 import * as AppsActions from './actions/appsActions';
 import * as AutoUpdateActions from './actions/autoUpdateActions';
 import * as ProxyActions from './actions/proxyActions';
-import * as SettingsAction from './actions/settingsActions';
+import * as UserDataActions from './actions/userDataActions';
 import '../../resources/css/launcher.scss';
 
 const config = remote.require('../main/config');
@@ -86,7 +86,7 @@ net.registerProxyLoginHandler((authInfo, callback) => {
 });
 
 render(rootElement, document.getElementById('webapp'), async () => {
-    await store.dispatch(SettingsAction.checkUserDataSetting(isSendingUserData));
+    await store.dispatch(UserDataActions.checkUserDataSetting(isSendingUserData));
     await store.dispatch(AppsActions.loadLocalApps());
     await store.dispatch(AppsActions.loadOfficialApps());
     await store.dispatch(AppsActions.setAppManagementShow());
