+++ src/launcher/reducers/settingsReducer.js
@@ -36,6 +36,7 @@
 
 import { Map, Record } from 'immutable';
 import * as SettingsActions from '../actions/settingsActions';
+import * as UserDataActions from '../actions/userDataActions';
 
 const InitialState = Record({
     shouldCheckForUpdatesAtStartup: true,
@@ -95,13 +96,13 @@ const reducer = (state = initialState, action) => {
             return state.set('isRemoveSourceDialogVisible', true).set('removeSource', action.name);
         case SettingsActions.SETTINGS_REMOVE_SOURCE_DIALOG_HIDE:
             return state.set('isRemoveSourceDialogVisible', false).set('removeSource', null);
-        case SettingsActions.SETTINGS_USER_DATA_DIALOG_SHOW:
+        case UserDataActions.SETTINGS_USER_DATA_DIALOG_SHOW:
             return state.set('isUserDataDialogVisible', true);
-        case SettingsActions.SETTINGS_USER_DATA_DIALOG_HIDE:
+        case UserDataActions.SETTINGS_USER_DATA_DIALOG_HIDE:
             return state.set('isUserDataDialogVisible', false);
-        case SettingsActions.SETTINGS_USER_DATA_SEND_ON:
+        case UserDataActions.SETTINGS_USER_DATA_SEND_ON:
             return state.set('isSendingUserData', true);
-        case SettingsActions.SETTINGS_USER_DATA_SEND_OFF:
+        case UserDataActions.SETTINGS_USER_DATA_SEND_OFF:
             return state.set('isSendingUserData', false);
         default:
             return state;
